+++ source/extensions/omni.isaac.merge_mesh/omni/isaac/merge_mesh/tests/__init__.py
@@ -7,4 +7,4 @@
 # license agreement from NVIDIA CORPORATION is strictly prohibited.
 #
 
-from .startup import *
+from .merge_mesh import *